Creating Dockerized test harness
