+++ aws/buckets.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+set -x
+awslocal s3 mb s3://testing
+set +x
