+++ ci/scripts/component.sh
@@ -1,0 +1,5 @@
+#!/bin/bash -eux
+
+pushd dp-upload-service
+  make docker-test-component
+popd
