+++ config/config_test.go
@@ -83,13 +83,6 @@ func TestSpec(t *testing.T) {
 				So(cfg.MinioSecretKey, ShouldEqual, "")
 				So(cfg.IsPublishing, ShouldBeTrue)
 				So(cfg.EncryptionDisabled, ShouldBeFalse)
-				So(cfg.EnableMongo, ShouldBeFalse)
-				So(cfg.MongoConfig.BindAddr, ShouldEqual, "localhost:27017")
-				So(cfg.MongoConfig.Collection, ShouldEqual, "")
-				So(cfg.MongoConfig.Database, ShouldEqual, "")
-				So(cfg.MongoConfig.Username, ShouldEqual, "")
-				So(cfg.MongoConfig.Password, ShouldEqual, "")
-				So(cfg.MongoConfig.IsSSL, ShouldBeFalse)
 			})
 		})
 	})
