+++ docker-compose.yml
@@ -1,0 +1,42 @@
+version: "3"
+
+services:
+  download-service:
+    image: template_test
+    ports:
+      - "23600:23600"
+    environment:
+      BIND_ADDR: ':23600'
+      AWS_REGION: 'eu-west1' # http://localstack:4572
+      BUCKET_NAME: 'testing'
+      ENCRYPTION_DISABLED: 'false'
+      GRACEFUL_SHUTDOWN_TIMEOUT: '5s'
+      HEALTHCHECK_INTERVAL: '5s'
+      HEALTHCHECK_CRITICAL_TIMEOUT: '5s'
+      VAULT_TOKEN: '0000-0000-0000-0000'
+      VAULT_ADDR: 'http://vault:8200'
+      VAULT_PATH: 'secret/shared/psk'
+
+  vault:
+    image: hashicorp/vault:latest
+    restart: always
+    ports:
+      - "8200:8200/tcp"
+    entrypoint: vault server -dev -dev-kv-v1 -dev-root-token-id="0000-0000-0000-0000"
+    environment:
+      VAULT_DEV_ROOT_TOKEN_ID: '0000-0000-0000-0000'
+      VAULT_DEV_LISTEN_ADDRESS: '0.0.0.0:8200'
+
+  localstack:
+    image: localstack/localstack
+    ports:
+      - '4566-4583'
+    environment:
+      - SERVICES=s3
+      - DEBUG=0
+      - AWS_DEFAULT_REGION=eu-west-1
+      - EDGE_PORT=4566
+    volumes:
+      - ./aws:/docker-entrypoint-initaws.d
+      - "${TMPDIR:-/tmp}/localstack:/tmp/localstack"
+      - "/var/run/docker.sock:/var/run/docker.sock"
