+++ features/health.feature
@@ -1,0 +1,6 @@
+Feature: Health check
+
+  Scenario:
+    When I GET "/health"
+    Then the HTTP status code should be "200"
+
