+++ service/moq_service_test.go
@@ -560,156 +560,6 @@ func (mock *HealthCheckerMock) StopCalls() []struct {
 	return calls
 }
 
-// Ensure, that MongoClientMock does implement service.MongoClient.
-// If this is not the case, regenerate this file with moq.
-var _ service.MongoClient = &MongoClientMock{}
-
-// MongoClientMock is a mock implementation of service.MongoClient.
-//
-// 	func TestSomethingThatUsesMongoClient(t *testing.T) {
-//
-// 		// make and configure a mocked service.MongoClient
-// 		mockedMongoClient := &MongoClientMock{
-// 			CheckerFunc: func(contextMoqParam context.Context, checkState *healthcheck.CheckState) error {
-// 				panic("mock out the Checker method")
-// 			},
-// 			CloseFunc: func(contextMoqParam context.Context) error {
-// 				panic("mock out the Close method")
-// 			},
-// 			URIFunc: func() string {
-// 				panic("mock out the URI method")
-// 			},
-// 		}
-//
-// 		// use mockedMongoClient in code that requires service.MongoClient
-// 		// and then make assertions.
-//
-// 	}
-type MongoClientMock struct {
-	// CheckerFunc mocks the Checker method.
-	CheckerFunc func(contextMoqParam context.Context, checkState *healthcheck.CheckState) error
-
-	// CloseFunc mocks the Close method.
-	CloseFunc func(contextMoqParam context.Context) error
-
-	// URIFunc mocks the URI method.
-	URIFunc func() string
-
-	// calls tracks calls to the methods.
-	calls struct {
-		// Checker holds details about calls to the Checker method.
-		Checker []struct {
-			// ContextMoqParam is the contextMoqParam argument value.
-			ContextMoqParam context.Context
-			// CheckState is the checkState argument value.
-			CheckState *healthcheck.CheckState
-		}
-		// Close holds details about calls to the Close method.
-		Close []struct {
-			// ContextMoqParam is the contextMoqParam argument value.
-			ContextMoqParam context.Context
-		}
-		// URI holds details about calls to the URI method.
-		URI []struct {
-		}
-	}
-	lockChecker sync.RWMutex
-	lockClose   sync.RWMutex
-	lockURI     sync.RWMutex
-}
-
-// Checker calls CheckerFunc.
-func (mock *MongoClientMock) Checker(contextMoqParam context.Context, checkState *healthcheck.CheckState) error {
-	if mock.CheckerFunc == nil {
-		panic("MongoClientMock.CheckerFunc: method is nil but MongoClient.Checker was just called")
-	}
-	callInfo := struct {
-		ContextMoqParam context.Context
-		CheckState      *healthcheck.CheckState
-	}{
-		ContextMoqParam: contextMoqParam,
-		CheckState:      checkState,
-	}
-	mock.lockChecker.Lock()
-	mock.calls.Checker = append(mock.calls.Checker, callInfo)
-	mock.lockChecker.Unlock()
-	return mock.CheckerFunc(contextMoqParam, checkState)
-}
-
-// CheckerCalls gets all the calls that were made to Checker.
-// Check the length with:
-//     len(mockedMongoClient.CheckerCalls())
-func (mock *MongoClientMock) CheckerCalls() []struct {
-	ContextMoqParam context.Context
-	CheckState      *healthcheck.CheckState
-} {
-	var calls []struct {
-		ContextMoqParam context.Context
-		CheckState      *healthcheck.CheckState
-	}
-	mock.lockChecker.RLock()
-	calls = mock.calls.Checker
-	mock.lockChecker.RUnlock()
-	return calls
-}
-
-// Close calls CloseFunc.
-func (mock *MongoClientMock) Close(contextMoqParam context.Context) error {
-	if mock.CloseFunc == nil {
-		panic("MongoClientMock.CloseFunc: method is nil but MongoClient.Close was just called")
-	}
-	callInfo := struct {
-		ContextMoqParam context.Context
-	}{
-		ContextMoqParam: contextMoqParam,
-	}
-	mock.lockClose.Lock()
-	mock.calls.Close = append(mock.calls.Close, callInfo)
-	mock.lockClose.Unlock()
-	return mock.CloseFunc(contextMoqParam)
-}
-
-// CloseCalls gets all the calls that were made to Close.
-// Check the length with:
-//     len(mockedMongoClient.CloseCalls())
-func (mock *MongoClientMock) CloseCalls() []struct {
-	ContextMoqParam context.Context
-} {
-	var calls []struct {
-		ContextMoqParam context.Context
-	}
-	mock.lockClose.RLock()
-	calls = mock.calls.Close
-	mock.lockClose.RUnlock()
-	return calls
-}
-
-// URI calls URIFunc.
-func (mock *MongoClientMock) URI() string {
-	if mock.URIFunc == nil {
-		panic("MongoClientMock.URIFunc: method is nil but MongoClient.URI was just called")
-	}
-	callInfo := struct {
-	}{}
-	mock.lockURI.Lock()
-	mock.calls.URI = append(mock.calls.URI, callInfo)
-	mock.lockURI.Unlock()
-	return mock.URIFunc()
-}
-
-// URICalls gets all the calls that were made to URI.
-// Check the length with:
-//     len(mockedMongoClient.URICalls())
-func (mock *MongoClientMock) URICalls() []struct {
-} {
-	var calls []struct {
-	}
-	mock.lockURI.RLock()
-	calls = mock.calls.URI
-	mock.lockURI.RUnlock()
-	return calls
-}
-
 // Ensure, that HTTPServerMock does implement service.HTTPServer.
 // If this is not the case, regenerate this file with moq.
 var _ service.HTTPServer = &HTTPServerMock{}
